refactor(quiz-service): consolidate Jest config and restructure tests
- Remove legacy jest.json and jest-e2e.json files
- Merge all Jest settings into jest.config.cjs (moduleFileExtensions, testRegex, transform, coverage, maxWorkers, etc.)
- Update root and package scripts: drop “test:e2e”, use --runInBand for tests and yarn test --coverage
- Relocate all e2e-spec files from “test/” into “src/…/controllers” and update import paths to point at test-utils
- Introduce “packages/quiz-service/test-utils” for shared mocks, bootstrap and asset data
- Update tsconfig.build.json to exclude test-utils instead of “test” directory
