+++ packages/quiz-service/jest.config.cjs
@@ -1,9 +1,20 @@
 // jest.config.cjs
 // eslint-disable-next-line no-undef
 module.exports = {
-  projects: ['<rootDir>/jest.json', '<rootDir>/test/jest-e2e.json'],
-  collectCoverage: true,
-  coverageDirectory: '<rootDir>/coverage',
+  moduleFileExtensions: ['js', 'json', 'ts'],
+  rootDir: 'src',
+  testRegex: '.*\\.(e2e-)?spec\\.ts$',
+  transform: {
+    '^.+\\.(t|j)s$': 'ts-jest',
+  },
+  moduleNameMapper: {
+    '^@quiz/common(.*)$': '<rootDir>/../../common/src$1',
+  },
+  collectCoverageFrom: ['**/*.(t|j)s'],
+  coverageDirectory: '../coverage',
+  testEnvironment: 'node',
+  transformIgnorePatterns: ['/node_modules/(?!@quiz/common)'],
   detectOpenHandles: true,
   forceExit: true,
+  maxWorkers: 1,
 }
