--- packages/quiz-service/jest.json
@@ -1,21 +1,0 @@
-{
-  "moduleFileExtensions": ["js", "json", "ts"],
-  "rootDir": "src",
-  "testRegex": ".*\\.spec\\.ts$",
-  "transform": {
-    "^.+\\.(t|j)s$": "ts-jest"
-  },
-  "moduleNameMapper": {
-    "^@quiz/common(.*)$": "<rootDir>/../../common/src$1"
-  },
-  "collectCoverageFrom": [
-    "**/*.(t|j)s"
-  ],
-  "coverageDirectory": "../coverage",
-  "testEnvironment": "node",
-  "transformIgnorePatterns": [
-    "/node_modules/(?!@quiz/common)"
-  ],
-  "detectOpenHandles": true,
-  "maxWorkers": 1
-}
