+++ packages/quiz-service/src/game/controllers/client-game.controller.e2e-spec.ts
@@ -4,11 +4,6 @@ import { GameMode, GameParticipantType, GameStatus } from '@quiz/common'
 import supertest from 'supertest'
 import { v4 as uuidv4 } from 'uuid'
 
-import { AuthService } from '../src/auth/services'
-import { Client, ClientModel } from '../src/client/services/models/schemas'
-import { Game, GameModel } from '../src/game/services/models/schemas'
-import { Player, PlayerModel } from '../src/player/services/models/schemas'
-
 import {
   createMockClientDocument,
   createMockGameDocument,
@@ -18,8 +13,15 @@ import {
   createMockQuestionTaskDocument,
   createMockQuitTaskDocument,
   offsetSeconds,
-} from './data'
-import { closeTestApp, createTestApp } from './utils/bootstrap'
+} from '../../../test-utils/data'
+import {
+  closeTestApp,
+  createTestApp,
+} from '../../../test-utils/utils/bootstrap'
+import { AuthService } from '../../auth/services'
+import { Client, ClientModel } from '../../client/services/models/schemas'
+import { Player, PlayerModel } from '../../player/services/models/schemas'
+import { Game, GameModel } from '../services/models/schemas'
 
 describe('ClientGameController (e2e)', () => {
   let app: INestApplication
