+++ packages/quiz-service/src/game/controllers/quiz-game.controller.e2e-spec.ts
@@ -17,11 +17,13 @@ import {
 import supertest from 'supertest'
 import { v4 as uuidv4 } from 'uuid'
 
-import { AuthService } from '../src/auth/services'
-import { ClientService } from '../src/client/services'
-import { QuizService } from '../src/quiz/services'
-
-import { closeTestApp, createTestApp } from './utils/bootstrap'
+import {
+  closeTestApp,
+  createTestApp,
+} from '../../../test-utils/utils/bootstrap'
+import { AuthService } from '../../auth/services'
+import { ClientService } from '../../client/services'
+import { QuizService } from '../../quiz/services'
 
 const multiChoiceQuestion: QuestionMultiChoiceDto = {
   type: QuestionType.MultiChoice,
