+++ packages/quiz-service/src/media/controllers/media.controller.e2e-spec.ts
@@ -5,9 +5,11 @@ import { INestApplication } from '@nestjs/common'
 import supertest from 'supertest'
 import { v4 as uuidv4 } from 'uuid'
 
-import { AuthService } from '../src/auth/services'
-
-import { closeTestApp, createTestApp } from './utils/bootstrap'
+import {
+  closeTestApp,
+  createTestApp,
+} from '../../../test-utils/utils/bootstrap'
+import { AuthService } from '../../auth/services'
 
 describe('MediaController (e2e)', () => {
   let app: INestApplication
@@ -73,7 +75,10 @@ describe('MediaController (e2e)', () => {
         await supertest(app.getHttpServer())
           .post('/api/media/uploads/photos')
           .set({ Authorization: `Bearer ${token}` })
-          .attach('file', join(__dirname, `./assets/photo.${extension}`))
+          .attach(
+            'file',
+            join(__dirname, `../../../test-utils/assets/photo.${extension}`),
+          )
           .expect(201)
           .expect((res) => {
             expect(res.body).toEqual({
@@ -82,7 +87,7 @@ describe('MediaController (e2e)', () => {
             return rm(
               join(
                 __dirname,
-                '../',
+                '../../../',
                 process.env.UPLOAD_DIRECTORY,
                 `/${dirname(res.body.filename)}`,
               ),
@@ -98,7 +103,7 @@ describe('MediaController (e2e)', () => {
       return supertest(app.getHttpServer())
         .post('/api/media/uploads/photos')
         .set({ Authorization: `Bearer ${token}` })
-        .attach('file', join(__dirname, './assets/empty.txt'))
+        .attach('file', join(__dirname, '../../../test-utils/assets/empty.txt'))
         .expect(422)
         .expect((res) => {
           expect(res.body).toEqual({
@@ -130,10 +135,10 @@ describe('MediaController (e2e)', () => {
 
       const photoId = uuidv4()
 
-      const srcFile = join(__dirname, `./assets/photo.webp`)
+      const srcFile = join(__dirname, `../../../test-utils/assets/photo.webp`)
       const dstFile = join(
         __dirname,
-        '../',
+        '../../../',
         process.env.UPLOAD_DIRECTORY,
         `/${clientId}/${photoId}.webp`,
       )
