+++ packages/quiz-service/test-utils/assets/photo.png
